feat: enhance TileHeader component with right button functionality and update usage in FAQCard and TipsAndTestimonial
